Main contents and community updated
