+++ pull_request_template.md
@@ -3,14 +3,17 @@ Wecode PR message template 및 체크 리스트 입니다.
 
 ## 수정 사항 간략한 한줄 요약
 (여기에 수정 사항에 대한 간략한 한줄 요약/제목 작성해 주세요. 예: 로그인 엔드포인트 구현)       
+1. 메인페이지 타일 react 컴포넌트 구현
+2. floating bar 추가
 
 ## 수정 사항들 자세한 내용
 (여기에 수정 사항에 대한 자세한 내용을 작성해 주세요)
-
+1. 메인페이지 타일 더미 데이터와 react 컴포넌트를 이용하여 구현
+2. floating bar hover는 아직 미구현
 
 
 ## 체크 리스트 (아래 사항들이 전부 체크되어야만 merge가 됩니다!)
-- [ ] 필요한 test들을 완료하였고 기능이 제대로 실행되는지 확인 하였습니다.
-- [ ] Wecode의 코드 스타일 가이드에 맞추어 코드를 작성 하였습니다.
-- [ ] 제가 의도한 파일들과 수정 사항들만 커밋이 된 것을 확인 하였습니다.
-- [ ] 본 수정 사항들을 팀원들과 사전에 상의하였고 팀원들 모두 해당 PR에 대하여 알고 있습니다.
+- [y] 필요한 test들을 완료하였고 기능이 제대로 실행되는지 확인 하였습니다.
+- [y] Wecode의 코드 스타일 가이드에 맞추어 코드를 작성 하였습니다.
+- [y] 제가 의도한 파일들과 수정 사항들만 커밋이 된 것을 확인 하였습니다.
+- [y] 본 수정 사항들을 팀원들과 사전에 상의하였고 팀원들 모두 해당 PR에 대하여 알고 있습니다.
