+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import { BrowserRouter as Router, Route} from "react-router-dom";
 
-import Banner from "./components/banner/Banner";
-import Main from "./pages/main/Main";
+import Banner from "./Components/Banner/Banner";
+import Main from "./Pages/Main/Main";
 
 class App extends Component {
     render() {
@@ -11,7 +11,6 @@ class App extends Component {
                 <Banner />
                 <Route exact path="/" component={Main} />
                 <Route path="/login" component />
-                <h1>App!</h1>
             </Router>
         );
     }
