+++ src/Components/Banner/Promo/Promo.scss
@@ -1,0 +1,12 @@
+.Promo {
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    width: 552px;
+    padding: 17px 18px;
+}
+
+.slide {
+    position: relative;
+    transition: linear 3s;
+}
