+++ src/Components/Banner/Promo/PromoItem/PromoItem.scss
@@ -1,0 +1,17 @@
+.PromoItem {
+    // position: absolute;
+    // top: 0;
+    width: 100%;
+    height: 100%;
+
+    .wrap{
+        width: 100%;
+        height: 100%;
+        position: relative;
+        span {
+            font-size: 14px;
+            line-height: 1.2;
+            font-family: "proxima nova", "proxima-nova", "Nanum Gothic", Helvetica, Arial, sans-serif;
+        }
+    }
+}
