+++ src/Components/FloatingBar/FloatingBar.scss
@@ -1,0 +1,47 @@
+.FloatingBar {
+    z-index: 1;
+    position: fixed;
+    right: 0;
+    bottom: 50px;
+    overflow: hidden;
+
+    .anchor {
+        height: 80px;
+        border-radius: 40px 0 0 40px;
+        padding: 0 10px 0 20px;
+        -webkit-transform: translateX(calc(100% - 70px));
+        transform: translateX(calc(100% - 70px));
+        -ms-transform: translateX(100%) translateX(-70px);
+        -webkit-transition: -webkit-transform 0.32s cubic-bezier(0.4, 0.9, 0.3, 1);
+        transition: -webkit-transform 0.32s cubic-bezier(0.4, 0.9, 0.3, 1);
+        transition: transform 0.32s cubic-bezier(0.4, 0.9, 0.3, 1);
+        transition: transform 0.32s cubic-bezier(0.4, 0.9, 0.3, 1), -webkit-transform 0.32s cubic-bezier(0.4, 0.9, 0.3, 1);
+        //768px 이상
+        display: flex;
+        align-items: center;
+        color: #ffffff;
+        background-color: #000000;
+
+        .icon {
+            margin-right: 10px;
+            font-size: 40px;
+            //768px 이상
+            display: inline-block;
+            width: 1em;
+            height: 1em;
+            fill: currentColor;
+        }
+
+        .content {
+            margin: 0;
+            font-size: 14px;
+            font-weight: 600;
+            line-height: 20px;
+            //768px 이상
+        }
+    }
+
+    .anchor:hover {
+        //on work
+    }
+}
