+++ src/Pages/Main/CommunityContainer/Communities/Communities.js
@@ -1,0 +1,29 @@
+import React, { Component } from 'react'
+
+import CommunityItem from "./CommunityItem/CommunityItem";
+
+import "./Communities.scss";
+
+class Communities extends Component {
+    render() {
+        const communities = this.props.Communities;
+        
+        return (
+            <div className="Communities">
+                <CommunityItem />
+                <CommunityItem />
+                <CommunityItem />
+                <CommunityItem />
+                <CommunityItem />
+                <CommunityItem />
+                {/* 더보기 생략
+                <div className="footer">
+
+                </div> 
+                */}
+            </div>
+        )
+    }
+}
+
+export default Communities
