+++ src/Pages/Main/CommunityContainer/Communities/Communities.scss
@@ -1,0 +1,5 @@
+.Communities {
+    overflow: hidden;
+    padding: 40px 0 60px;
+    text-align: left;
+}
