+++ src/Pages/Main/CommunityContainer/Communities/CommunityItem/CommunityItem.scss
@@ -1,0 +1,114 @@
+.CommunityItem {
+    display: inline-block;
+    //display: inline-block  ,for flex
+    width: 25%;
+    height: auto;
+    padding: 5px;
+    //animation ?
+    backface-visibility: hidden;
+    -webkit-backface-visibility: hidden;
+    -webkit-transform-style: preserve-3d;
+    transform-style: preserve-3d;
+    -webkit-perspective: 1200px;
+    overflow: hidden;
+
+    .item-wrapper {
+        position: relative;
+        width: 100%;
+        height: 24vw;
+        // max-height: 22vw;
+        cursor: pointer;
+        -webkit-transform-style: preserve-3d;
+        transform-style: preserve-3d;
+        -webkit-transform: translate3d(0,0,0) rotateX(0) rotateY(0);
+        transform: translate3d(0,0,0) rotateX(0) rotateY(0);
+        -webkit-transition: all .8s ease;
+        -o-transition: all .8s ease;
+        transition: all .8s ease;
+        -webkit-backface-visibility: hidden;
+        backface-visibility: hidden;
+        background: url(https://cdn.attractt.com/embed/images/error/404-img-320.png) no-repeat;
+        background-position: center;
+        -webkit-background-size: contain;
+        background-size: contain;
+        opacity: 1;
+        
+            .thumbnail {
+                display: block;
+                width: 100%;
+                height: 100%;
+                -o-object-fit: cover;
+                object-fit: fill;
+                background-color: #fff;
+            }
+        
+            .overlay {
+                position: absolute;
+                left: 0;
+                bottom: 0;
+                width: 100%;
+                height: 100%;
+        
+                .user {
+                    position: absolute;
+                    z-index: 200;
+                    left: 0;
+                    bottom: 0;
+                    width: 100%;
+                    height: 50px;
+                    padding: 10px;
+                    -webkit-box-sizing: border-box;
+                    -moz-box-sizing: border-box;
+                    box-sizing: border-box;
+        
+                    .profile {
+                        position: relative;
+                        width: 30px;
+                        height: 30px;
+                        border-radius: 15px;
+                        z-index: 200;
+                    }
+        
+                    .info {
+                        position: absolute;
+                        top: 11px;
+                        left: 53px;
+                        z-index: 200;
+        
+                        .user-name {
+                            display: block;
+                            color: #fff;
+                            font-size: 14px;
+                        }
+                        .time {
+                            display: block;
+                            color: #fff;
+                            font-size: 12px;
+                            font-weight: 100;
+                        }
+                    }
+
+                    .tag-icon {
+                        position: absolute;
+                        bottom: 12px;
+                        right: 12px;
+                        width: 16px;
+                        height: 16px;
+                        z-index: 300;
+
+                        i {
+                            display: inline-block;
+                            width: 100%;
+                            height: 100%;
+                            background: url(https://cdn.attractt.com/www/images/sprite/new/sprite_tag.png) no-repeat;
+                            background-position: -20px -20px!important;
+                            -webkit-background-size: 80px 263px!important;
+                            background-size: 80px 263px!important;
+                        }
+                    }
+                }
+            }
+
+
+    }
+}
