+++ src/Pages/Main/CommunityContainer/CommunityContainer.scss
@@ -1,0 +1,20 @@
+.CommunityContainer {
+    position: relative;
+    margin: 0 auto;
+    overflow: hidden;
+    max-width: 1440px;
+
+    .community-wrapper {
+        margin: 0 auto; 
+        padding: 60px 0; // 768px 이상
+        text-align: center; // ""
+
+        .title {
+            margin: 0; // 768px 이상
+            font-size: 36px; // ""
+            line-height: 1; // ""
+            font-weight: 700;
+            text-transform: uppercase;
+        }
+    }
+}
