+++ src/Pages/Main/ContentsContainer/ContentNav/ContentNav.scss
@@ -1,0 +1,67 @@
+.ContentNav {
+    width: 100%;
+    padding: 60px 0;
+    text-align: center;
+    background-color: #ffffff;
+
+    .title {
+        font-weight: 700;
+        font-size: 36px; //768px 이상
+        line-height: 1; //768px 이상
+    }
+
+    .link {
+        display: inline-flex;
+        margin: 20px 0 0;
+        font-size: 16px;
+        font-weight: 600;
+        line-height: 1;
+
+        .icon {
+            display: inline-block;
+            position: relative;
+            height: 1em; //768px 이상
+            width: 1em;
+            fill: currentColor;
+            max-width: 100%;
+            vertical-align: middle;
+        }
+
+        .link-text {
+            margin-left: 9px; //768px 이상
+            line-height: 0.9;
+        }
+    }
+
+    .list {
+        display: flex;
+        margin: 20px auto 0; //768px 이상
+        width: 100%; //768px 이상
+        -webkit-user-select: none;
+        -moz-user-select: none;
+        -ms-user-select: none;
+        user-select: none;
+        -webkit-touch-callout: none;
+        -khtml-user-select: none;
+        -ms-touch-action: pan-y;
+        touch-action: pan-y;
+        -webkit-tap-highlight-color: transparent;
+        //inner div
+        overflow: hidden;
+        //margin: 0; padding: 0;
+
+        .items {
+            display: flex;
+            justify-content: space-around;
+
+            position: relative;
+            top: 0;
+            left: 0;
+            margin: 0 auto;
+            width: 50%;
+            opacity: 1;
+            transform: translate3d(0px, 0px, 0px);
+        }
+    }
+
+}
