+++ src/Pages/Main/ContentsContainer/Contents/Contents.scss
@@ -1,0 +1,6 @@
+.Contents {
+    display: inline-block;
+    width: 50%;
+    height: 716px;
+    vertical-align: middle; // important!!!
+}
