+++ src/Pages/Main/ContentsContainer/ContentsContainer.scss
@@ -1,0 +1,5 @@
+.ContentsContainer {
+    margin: 0 auto;
+    max-width: 1440px;
+    font-size: 0;
+}
