+++ src/Styles/Common.scss
@@ -1,0 +1,39 @@
+* {
+    box-sizing: border-box;
+    text-decoration: none;
+    color: black;
+    font-family: "proxima nova", "proxima-nova", "Nanum Gothic", Helvetica, Arial, sans-serif;
+}
+
+img {
+    max-width: 100%;
+    max-height: 100%;
+    border-style: none;
+    border: 0;
+    height: auto;
+    vertical-align: middle;
+}
+
+a {
+    color: inherit;
+    text-decoration: none;
+    background-color: transparent;
+    cursor: pointer;
+}
+
+input {
+    box-shadow: none;
+}
+
+input[type="text"], input[type="password"], textarea, select { 
+    outline: none;
+}
+
+button {
+    border: 0;
+    outline: 0;
+}
+
+button:hover {
+    cursor: pointer;
+}
