--- src/components/banner/Promo/Promo.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-
-import PromoItem from "./PromoItem";
-
-class Promo extends Component {
-    render() {
-        return (
-            <div className="promo">
-                <PromoItem />
-            </div>
-        )
-    }
-}
-
-export default Promo;
