--- src/components/banner/Promo/PromoItem.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react'
-
-class PromoItem extends Component {
-    render() {
-        return (
-            <div className="promoItem">
-                <span>
-                    NOTICE
-                    <br/>
-                    신규 가입 1만원 할인 코드 사용 관련
-                </span>
-            </div>
-        )
-    }
-}
-
-export default PromoItem;
