+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
-import "./styles/Reset.scss";
-import "./styles/Common.scss";
+import "./Styles/Reset.scss";
+import "./Styles/Common.scss";
 
 ReactDOM.render(
   <React.StrictMode>
