--- src/pages/main/ContentsContainer/Contents/ContentForOne.scss
@@ -1,92 +1,0 @@
-.ContentForOne {
-    display: inline-block;
-    position: relative;
-    width: 50%;
-    // min-width: 25%;
-    // width: 286px;
-    height: 358px;
-    background-color: skyblue;
-
-    .product-wish {
-        position: absolute;
-        top: 2.0833vw;  // 768px 이하 시 30px or 14px
-        left: 2.0833vw; // 768px 이하 시 30px or 14px
-        width: 44px;
-        height: 44px;
-        font-size: 18px;
-        color: #000000;
-        -webkit-transform: scale(1);
-        transform: scale(1);
-        -webkit-transition: -webkit-transform 213ms cubic-bezier(0.4, 0.9, 0.3, 2.25);
-        transition: -webkit-transform 213ms cubic-bezier(0.4, 0.9, 0.3, 2.25);
-        transition: transform 213ms cubic-bezier(0.4, 0.9, 0.3, 2.25);
-        transition: transform 213ms cubic-bezier(0.4, 0.9, 0.3, 2.25), -webkit-transform 213ms cubic-bezier(0.4, 0.9, 0.3, 2.25);
-    
-
-        .wish-action {
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            .remove {
-                -webkit-transform: scale(1);
-                transform: scale(1);
-                -webkit-transition: -webkit-transform 213ms ease-in-out;
-                transition: -webkit-transform 213ms ease-in-out;
-                transition: transform 213ms ease-in-out;
-                transition: transform 213ms ease-in-out,-webkit-transform 213ms ease-in-out;
-            }
-
-            .remove:not(.self-active) {
-                -webkit-transform: scale(0);
-                transform: scale(0);
-            }
-        }
-    }
-
-    .product-badges {
-        position: absolute;
-        top: 2.0833vw;  // 768px 이하 시 30 or 14 px
-        right: 2.0833vw; // 768px 이하 시 30 or 14 px
-
-        display: flex;
-        flex-direction: column;
-        align-items: flex-end;
-
-        .badge-text {
-            display: -webkit-inline-box;
-            display: -webkit-inline-flex;
-            display: -ms-inline-flexbox;
-            display: inline-flex;
-            font-size: 14px;
-            line-height: 1.2;
-            text-transform: uppercase;
-            font-weight: 700;
-        }
-    }
-
-    .product-detail {
-        position: absolute;
-        bottom: 2.0833vw;   // 768px 이하 시 30px
-        left: 2.0833vw;     // 768px 이하 시 30px
-        max-width: calc( 100% - 4.16666vw); // 768 이하 시 calc(100% - 60px)
-        line-height: 1.3333em;
-        font-size: 14px;
-        letter-spacing: -.025em;
-        word-break: keep-all;
-        word-wrap: break-word;
-        overflow-wrap: break-word;
-
-        .price {
-            margin-top: 4px;
-        }
-    }
-
-
-}
