--- src/pages/main/ContentsContainer/Contents/ContentForTwo.scss
@@ -1,113 +1,0 @@
-.ContentForTwo {
-    position: relative;
-    font-size: 14px;
-    line-height: 1.3333em;
-    letter-spacing: -.025em;
-    word-break: keep-all;
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-    overflow: hidden;
-    display: inline-block;
-    width: 100%;
-    // min-width: 50%;
-
-    // width: 574px;
-    height: 358px;
-
-    background-color: springgreen;
-
-    .product {
-        position: absolute;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        margin: auto;
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-        font-family: "object-fit: cover;";
-
-        .product-img {
-            display: block;
-            width: 100%;
-            height: 100%;
-            object-fit: cover;
-            font-family: "object-fit: cover;";
-        }
-    }
-
-    .tile-overlay {
-        padding: 2.0833vw; 
-        background-color: transparent; //768px 이상
-        position: absolute;
-        top: 0;
-        left: 0;
-        // padding: 30px;
-        width: 100%;
-        height: 100%;
-        pointer-events: none;
-        -webkit-box-orient: vertical;
-        -webkit-box-direction: normal;
-        -webkit-flex-direction: column;
-        -ms-flex-direction: column;
-        flex-direction: column;
-        //flex
-        display: flex;
-        //justify-content
-        -webkit-box-pack: start;
-        -webkit-justify-content: flex-start;
-        -ms-flex-pack: start;
-        justify-content: flex-start;
-        //text-align
-        text-align: left;
-
-        p {
-            margin-bottom: .9722vw;
-            margin: 0 0 14px;    
-        }
-
-        .badge {
-            line-height: 1.375em;
-            text-transform: uppercase;
-            font-size: calc(12px + 0.25*(100vw - 767px)/42.0625);//768이하 12px
-        }
-
-        .title {
-            line-height: 1.2222em;
-            text-transform: uppercase;
-            // font-size: 30px;
-            // font-size: calc(30px + 1.25*(100vw - 767px)/42.0625); //768이하 30px, 1440이상 50px
-            font-size: 50px;
-            font-weight: 700;
-        }
-
-        .description {
-            line-height: 1.375em;
-            font-size: calc(12px + 0.25*(100vw - 767px)/42.0625); //768이하 12px
-        }
-
-        .link {
-            display: inline-flex;
-
-            .icon {
-                display: inline-block;
-                position: relative;
-                height: 1.125em;
-                width: 1.125em;
-                top: .125em;
-                fill: currentColor;
-                max-width: 100%;
-                vertical-align: middle;
-                line-height: 1.375em;
-            }
-
-            .link-text {
-                margin-left: .5625em;
-                font-weight: 600;
-                line-height: 1.375em;
-            }
-        }
-
-    }
-}
