--- src/pages/main/ContentsContainer/Contents/Contents.scss
@@ -1,4 +1,0 @@
-.Contents {
-    display: inline-block;
-    width: 50%;
-}
