--- src/pages/main/ContentsContainer/ContentsContainer.scss
@@ -1,4 +1,0 @@
-.ContentsContainer {
-    margin: 0 auto;
-    max-width: 1440px;
-}
