--- src/styles/Common.scss
@@ -1,30 +1,0 @@
-* {
-    box-sizing: border-box;
-    text-decoration: none;
-    color: black;
-    font-family: "proxima nova", "proxima-nova", "Nanum Gothic", Helvetica, Arial, sans-serif;
-}
-
-a {
-    color: inherit;
-    text-decoration: none;
-    background-color: transparent;
-    cursor: pointer;
-}
-
-input {
-    box-shadow: none;
-}
-
-input[type="text"], input[type="password"], textarea, select { 
-    outline: none;
-}
-
-button {
-    border: 0;
-    outline: 0;
-}
-
-button:hover {
-    cursor: pointer;
-}
